@@ -1,71 +1,71 @@
 ﻿using Books.Models;
 using Books.Services;
+
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
-namespace Books.Data
+namespace Books.Data;
+
+public class BooksContext : DbContext, IBookChapterService
 {
-    public class BooksContext : DbContext, IBookChapterService
+    public BooksContext(DbContextOptions<BooksContext> options)
+        : base(options)
     {
-        public BooksContext(DbContextOptions<BooksContext> options)
-            : base(options)
-        {
-        }
+    }
 
-        public DbSet<BookChapter> Chapters => Set<BookChapter>();
+    public DbSet<BookChapter> Chapters => Set<BookChapter>();
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<BookChapter>().Property(b => b.Title).HasMaxLength(80);
-        }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<BookChapter>().Property(b => b.Title).HasMaxLength(80);
+    }
 
-        public async Task<BookChapter> AddAsync(BookChapter chapter)
+    public async Task<BookChapter> AddAsync(BookChapter chapter)
+    {
+        await Chapters.AddAsync(chapter);
+        await SaveChangesAsync();
+        return chapter;
+    }
+
+    public async Task AddRangeAsync(IEnumerable<BookChapter> chapters)
+    {
+        try
         {
-            await Chapters.AddAsync(chapter);
+            await this.Chapters.AddRangeAsync(chapters);
             await SaveChangesAsync();
-            return chapter;
         }
-
-        public async Task AddRangeAsync(IEnumerable<BookChapter> chapters)
+        catch (Exception ex)
         {
-            try
-            {
-                await this.Chapters.AddRangeAsync(chapters);
-                await SaveChangesAsync();
-            }
-            catch (Exception ex)
-            {
-                throw;
-            }
+            throw;
         }
+    }
 
-        public async Task<IEnumerable<BookChapter>> GetAllAsync()
-        {
-            var chapters = await Chapters.ToListAsync();
-            return chapters;
-        }
+    public async Task<IEnumerable<BookChapter>> GetAllAsync()
+    {
+        var chapters = await Chapters.ToListAsync();
+        return chapters;
+    }
 
-        public async Task<BookChapter?> FindAsync(Guid id)
-        {
-            var chapter = await Chapters.FindAsync(id);
-            return chapter;
-        }
+    public async Task<BookChapter?> FindAsync(Guid id)
+    {
+        var chapter = await Chapters.FindAsync(id);
+        return chapter;
+    }
 
-        public async Task<BookChapter?> RemoveAsync(Guid id)
+    public async Task<BookChapter?> RemoveAsync(Guid id)
+    {
+        var chapter = await Chapters.FindAsync(id);
+        if (chapter is not null)
         {
-            var chapter = await Chapters.FindAsync(id);
             Chapters.Remove(chapter);
             await SaveChangesAsync();
-            return chapter;
         }
+        return chapter;
+    }
 
-        public async Task<BookChapter?> UpdateAsync(BookChapter chapter)
-        {
-            Chapters.Update(chapter);
-            await SaveChangesAsync();
-            return chapter;
-        }
+    public async Task<BookChapter?> UpdateAsync(BookChapter chapter)
+    {
+        Chapters.Update(chapter);
+        await SaveChangesAsync();
+        return chapter;
     }
 }
