@@ -1,18 +1,13 @@
 ﻿using Books.Models;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 
-namespace Books.Services
+namespace Books.Services;
+
+public interface IBookChapterService
 {
-    public interface IBookChapterService
-    {
-        Task<BookChapter> AddAsync(BookChapter chapter);
-        Task AddRangeAsync(IEnumerable<BookChapter> chapters);
-        Task<IEnumerable<BookChapter>> GetAllAsync();
-        Task<BookChapter?> FindAsync(Guid id);
-        Task<BookChapter?> RemoveAsync(Guid id);
-        Task<BookChapter?> UpdateAsync(BookChapter chapter);
-    }
+    Task<BookChapter> AddAsync(BookChapter chapter);
+    Task AddRangeAsync(IEnumerable<BookChapter> chapters);
+    Task<IEnumerable<BookChapter>> GetAllAsync();
+    Task<BookChapter?> FindAsync(Guid id);
+    Task<BookChapter?> RemoveAsync(Guid id);
+    Task<BookChapter?> UpdateAsync(BookChapter chapter);
 }
