@@ -1,6 +1,5 @@
 ﻿using System;
 
-namespace Books.Models
-{
-    public record BookChapter(Guid Id, int Number, string Title, int PageCount);
-}
+namespace Books.Models;
+
+public record BookChapter(Guid Id, int Number, string Title, int PageCount);
