@@ -1,22 +1,29 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using Books.Data;
+using Books.Services;
 
-namespace GRPCService
+using GRPCService.Services;
+
+using Microsoft.EntityFrameworkCore;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Additional configuration is required to successfully run gRPC on macOS.
+// For instructions on how to configure Kestrel and gRPC clients on macOS, visit https://go.microsoft.com/fwlink/?linkid=2099682
+
+// Add services to the container.
+builder.Services.AddGrpc();
+builder.Services.AddDbContext<IBookChapterService, BooksContext>(options =>
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+    string connectionString = builder.Configuration.GetConnectionString("BooksConnection");
+    if (connectionString is null) throw new InvalidOperationException("Configure the connection string");
+    options.UseSqlServer(connectionString);
+});
 
-        // Additional configuration is required to successfully run gRPC on macOS.
-        // For instructions on how to configure Kestrel and gRPC clients on macOS, visit https://go.microsoft.com/fwlink/?linkid=2099682
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
-}
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+app.MapGrpcService<BooksService>();
+app.MapGrpcService<SensorService>();
+app.MapGet("/", () => "Use a gRPC client!");
+
+app.Run();
